+++ pkg/bhlquest.go
@@ -6,24 +6,29 @@ import (
 	"github.com/gnames/bhlquest/pkg/config"
 	"github.com/gnames/bhlquest/pkg/ent/bhln"
 	"github.com/gnames/bhlquest/pkg/ent/embed"
+	"github.com/gnames/bhlquest/pkg/ent/llmutil"
 )
 
 type bhlquest struct {
 	cfg  config.Config
 	bhln bhln.BHLN
 	emb  embed.Embed
+	llm  llmutil.LlmUtil
 }
 
 func New(
 	cfg config.Config,
 	bhln bhln.BHLN,
 	emb embed.Embed,
+	llm llmutil.LlmUtil,
 ) BHLQuest {
 	res := bhlquest{
 		cfg:  cfg,
 		bhln: bhln,
 		emb:  emb,
+		llm:  llm,
 	}
+
 	return res
 }
 
@@ -38,6 +43,8 @@ func (bq bhlquest) Init() error {
 	if err != nil {
 		return err
 	}
+
+	err = bq.emb.Populate(ids)
 
 	_ = ids
 	return nil