+++ pkg/ent/llmutil/interface.go
@@ -3,5 +3,5 @@ package llmutil
 import "github.com/gnames/bhlquest/pkg/ent/embed"
 
 type LlmUtil interface {
-	Embed([]embed.Chunk) []embed.Chunk
+	Embed([]embed.Chunk) ([]embed.Chunk, error)
 }