+++ cmd/flags.go
@@ -6,6 +6,7 @@ import (
 	"os"
 
 	bhlquest "github.com/gnames/bhlquest/pkg"
+	"github.com/gnames/bhlquest/pkg/config"
 	"github.com/spf13/cobra"
 )
 
@@ -21,11 +22,22 @@ func debugFlag(cmd *cobra.Command) {
 	}
 }
 
+func portFlag(cmd *cobra.Command) {
+	i, _ := cmd.Flags().GetInt("port")
+	if i > 0 {
+		opts = append(opts, config.OptPort(i))
+	}
+}
+
 func versionFlag(cmd *cobra.Command) {
 	b, _ := cmd.Flags().GetBool("version")
 	if b {
 		version := bhlquest.GetVersion()
-		fmt.Print(version)
+		fmt.Printf(
+			"\nVersion: %s\nBuild:   %s\n\n",
+			version.Version,
+			version.Build,
+		)
 		os.Exit(0)
 	}
 }