+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added v0.0.1
 
+- Connect to llmutil RESTful service (#4)
 - Create database reset with pgvector for the data (#3)
 - Return list of item ids (#2)